feat. add recent 3 records in tableview

+++ ZzzBank/Domain/Entities/RepayRecord.swift
@@ -10,7 +10,7 @@ import RealmSwift
 
 final class RepayRecord: Object, DateSortable {
     @objc dynamic var id = UUID().uuidString
-    @objc dynamic var repayTime: Double = 0.0
+    @objc dynamic var repayTime: Int = 0
     @objc dynamic var date: Date = Date()
     
     override static func primaryKey() -> String? {
